-Wsign-compare

+++ Bugsnag/Breadcrumbs/BugsnagBreadcrumbs.m
@@ -224,7 +224,7 @@ void BugsnagBreadcrumbsWriteCrashReport(const BSG_KSCrashReportWriter *writer) {
     writer->beginArray(writer, "breadcrumbs");
     for (unsigned int i = g_context.firstFileNumber; i < g_context.nextFileNumber; i++) {
         int result = snprintf(path, sizeof(path), "%s/%u.json", g_context.directoryPath, i);
-        if (result < 0 || result >= sizeof(path)) {
+        if (result < 0 || result >= (int)sizeof(path)) {
             bsg_log_err(@"Breadcrumb path is too long");
             continue;
         }
